Address layout bug affecting firefox, bug #1821.
Fix a layout issue that impacts at least some versions of firefox,
described in http://redmine/webtoolkit.eu/issues/1821.

It appears that the scrollHeight in Firefox is not being updated
as expected when there is no scroller and a height is set.
As a result, divs are not resized as expected.

This update removes the height constraint prior to measuring the
preferred height.  It may not be an optimal solution.

Note: this update addresses both height and width, although the
comments reflect a focus on height.  This code has not been extensively
tested, and should be reviewed and tested more completely to insure that
it does not break existing applications.