Fix for queued browser event display errors #1858.
When pending events are queued on the browser (due to communication or
server delay, etc.), they are are processed in a batch on the server.
The changes are currently accumulated in a way that the intent
is not always preserved.  Because duplicates are allowed to accumulate,
it appears that added style classes can be superceded by removed style
classes earlier in the batch sequence.

This fix attempts to address the issue by only retaining the latest
add / remove requested for each style class.

At least that's the intent....  Please review and test!