Hangman example: disconnect globalKeyPressed on LettersWidget destruct
This also avoids a possible segfault introduced by our signals +
std::bind not being automatically disconnected when "this" is deleted.

With Boost.Signals2 and boost::bind() this was not a problem, because
widgets were trackable. Another way to fix this would be use bindSafe,
or by using observing_ptr, but disconnecting the slots is a bit nicer.