Improved performance of adding rows and columns to the WTable.
The issue was the calling of WTableRow::rowNum() for every added cell,
which iterated over the whole table.

The methods WTableRow::createCell() and WTableRow::insertColumn() got
the new row argument defaulting to -1 that WTable can use when
creating new elements with WTable::insertRow() and
WTable::insertColumn(). These methods are called by WTable::expand()
so the performance improvement affects every operation that expands
the table.

The performance of populating a new table row by row is increased
with this commit from O(n^2+n) to O(n).