Fix issue #6260
This reverts some earlier change from 2010-10-18 to WTreeView:

EventSignal::connect(): added an overload which accepts a JavaScript
function, as a short-cut for JSlot when connection management is not
needed.

Evidently, connection management was indeed needed in this case.

Also made sure that slots are updated to the new id by doing a re-setup
if the object name or id is set after construction.