Fix WebSocket issues #7034 and #7039
Per message deflate is only indicated in the first frame of the message,
so we need to remember it instead of overwriting frameCompressed_ on
every frame (patch for issue #7034 by Bruce Toll)

Issue #7039: Sometimes it happens that the end of a frame and the start
of another end up in the same call to parseWebSocketMessage. We need to
make sure that we don't discard the first frame, and actually get out of
the loop, inflate it (if using PMD), and give it to the reply to
consume.

We'll loop over parseWebSocketMessage in parseBody as long as we still
have data and the request state is partial.

Based on the patch for issue #7039 by Bruce Toll.