Use unsigned int for ackId
signed integer overflow is undefined behaviour, and WRandom::get()
actually produces an unsigned integer. That random integer may be close
enough to INT_MAX or INT_MIN that it actually matters, and overflow is
possible.

Also simplified the expectedAckId_ - updateId < 5 check. We can just
rely on the well defined unsigned integer overflow behaviour. It's not
necessary to check whether updateId < expectedAckId_.

Added operator<< (unsigned int) to WStringStream so we don't need to
cast it to long long.