Use implementationIsFlexLayout() to determine whether flex item impl should be created
This is a fix for this issue:
https://redmine.webtoolkit.eu/boards/2/topics/16551

This fix does not cause a regression (WTreeViews with a fixed size would
not be properly confined).

The issue was that StdWidgetItemImpls would end up in a FlexLayoutImpl

This still does not fix issue #7230, where a layout implementation is
changed to a different one due to a setResizable being enabled (not
supported for flex implementation).