Added tests for WTable
Two of them (moveRow2 and moveColumn2) do not pass because of a bug in
the WTable's code, which is not adressed in this commit. That's why
these ones are disabled.

The tests are not that comprehensive and were created to test the
correctnes of code in future commits adressing WTable's performance.