Improved performance of adding big number of rows to WTable
- Added new field to WTableRow - rowNumber_ - to track its position in
  the table_.
- Added WTable::updateRowNumbers() function to update row numbers for
  rows that need updating.
- Modified WTable::insertRow(), WTable::removeRow() and
  WTable::moveRow() to update row numbers for the rows affected by
  these operations.
- Modified WTableCell::rowNum() to use WTableCell::rowNumber_ if
  present.

Adding rows at the end of the array is now much, much faster. It
doesn't need to iterate over every cell anymore.
Modified methods from WTable are a little slower than before when
removing, inserting or moving rows from the middle of the array. This
is because of added overhead of having to update row numbers but this
slowdown is negligible.