Use GCC/clang visibility everywhere
Setting the default of Wt libs to hidden using the CMake target
properties "CXX_VISIBILITY_PRESET" and "VISIBILITY_INLINES_HIDDEN".

Also making sure that DATE_BUILD_DLL is set for date library.

WServer has a special visibility story. Some of its symbols are in the
wt library, some of them in the connector library. On Windows, the
visibility does not have to be specified for the class itself, only the
members. However, on non-Windows, we need it because WServer has dynamic
type info (virtual destructor), so we add WTCONNECTOR_API.

We don't need the __GNUC__ >= 4 check, because Wt 4 needs at least GCC
4.9. Also, in some cases the ..._API defines were still empty on
non-Windows, so we made sure that the visibility was defined even on
those platforms.

There was something not completely right about how the compile defs for
tz.cpp were being set, so some of the compile defs wouldn't take.
Rewrote that a bit. It needs DATE_BUILD_DLL on Linux.

These visibility tweaks were inspired by the "Deep CMake for Library
Authors" talk by Craig Scott at CppCon 2019