Add wt-config-version.cmake
Following advice from "Deep CMake for Library Authors" talk, I added a
version file to the Wt CMake package.

It matches the version just as precisely as it was specified, e.g.

Wt 4.1.1 is matched by 4, 4.1, 4.1.1, and 4.1.1.whatever, with 4.1.1
being an exact match

Examples of non-matches:
Wt 4.1.1 is not matched by 3, 4.2, or 4.1.0

The autogenerated version.cmake would also do a bitness check. Let's
also add that check.