Issue #7210: fixing removeItem with JS layouts
If you remove an item from a JS layout, the JS layout will do the
remove, and rendering the remove is no longer necessary.

However, there are special widgets that have a special remove (like
WTextEdit), and that is simply not performed when in a JS layout (a
known other issue).