Revert "Don't do const_cast and then modify!"
This reverts part of commit 6095a0f327e76df6d21da89482a6c544245622dd.

The problem here is that insertions into a std::set will cause
it to do comparisons, but operator< of WModelIndex uses parent().

WSortFilterProxyModel treats internalPointer() as if it is a pointer
to WSortFilterProxyModel::Item, but toRawIndex() will turn it into
something else (in the case of WStandardItemModel this is a
WStandardItem pointer). This causes segfaults.

This code is still UB, but at least it doesn't currently crash. We'll
need to find a better solution, though.