Alleviate performance issues when expanding entire treeview
isExpanded() does a find in the expandedSet_, which is relatively slow
because of the pretty expensive operator< of WModelIndex (topographical
order).

The performance hit becomes worse when a WSortFilterProxyModel is used,
since you get a WModelIndex find in the expandedSet_ and a WModelIndex
find in the mapping, which uses std::map.

The expandedSet_ now uses an unordered_set, only ordering the set when
necessary for shiftModelIndexes. (Can we also do shiftModelIndexes more
efficiently?)

Added encodeAsRawIndexes/decodeFromRawIndexes taking an unordered_Set to
WModelIndex to support this.