WTreeView: faster shiftModelIndexes for unordered_set
Instead of sorting the set and then using lower_bound, let's just go
over all of the indexes after "start" and check if they are in the set.
This turns out to be faster than when using std::set in many cases
(unless the set is very small or the model extremely large). Note that
using a std::set with topological ordering also adds a large cost to
other operations, so performance should usually be better overall when
using an unordered set.

Added unit tests for tree view (shifting expanded model indexes)