Delay lost password token removal until the password was changed
This is a behavioral change, so it will not be included in 4.2.1.

Before, if a user clicked on a lost password link, it would immediately
invalidate that token. If the user then did not use that link to update
their password, the token could no longer be used, and the user would
have to go through the lost password procedure again.

Now, the token will stay valid until the password is actually changed.
Also, if the password is changed through other means, the update
password token is invalidated (e.g. when an administrator changes the
password using some custom code).