WWebWidget: move id_ out of OtherImpl
This fixes an issue when setId() is used.

The destructor of jsScrollVisibilityChanged_ (a JSignal) in OtherImpl
uses WWebWidget::id() to unexpose the signal. However, resetting the
otherImpl_ causes it to be set to nullptr, so the WObject's id is used
instead of the id set in OtherImpl.

Simply moving this id_ out of OtherImpl and making sure it gets
destroyed after OtherImpl solves this problem.