Fix issue #7518: make password validator result depend on isMandatory()
The default implementation simply runs the password through the
validation algorithm and returns Valid/Invalid. It should actually
return InvalidEmpty if the field was left empty by default (hence
setMandatory(true) in the ctor), or it should return Valid if left empty
and the user has used setMandatory(false), and not run it through the
validation algorithm.