Make sure Google Maps is loaded only once
If more than one instance of WGoogleMap was created in a WApplication,
we'd get a JavaScript warning, and at least one of the maps wouldn't
work correctly.

Added a bit of JavaScript that instead checks whether Google Maps is
already loaded (can do callback immediately), or is already being loaded
(can add to list of callbacks).

Note: all instances of WGoogleMap in a WApplication should use the same
API version!