Connect loginChanged() earlier
If the user connected a slot to loginChanged(), and that slot uses the
session id (e.g. generating a resource url()), then that session id
would become wrong after onLoginChange (which causes the session id to
be changed). So let's connect it first, so it has a better chance of
going first.

Note: sadly not a complete fix, the user may still connect their slot
before creating the AuthWidget.