Issue #7617: use std::shared_ptr for ApplicationEvent
One of bindSafe's intended uses is to safely pass the pointer to an
observable to WServer::post. However, copying an observing_ptr is not
threadsafe. queueEvent would cause a copy of the callback (and thus the
observing_ptr), and this callback would also be destroyed without the
UpdateLock.

Destruction of an observing_ptr should be safe when the WApplication is
already gone, since then the observing_ptr will already be nulled out
(making destruction a no-op).