Issue #7669: fixing WSuggestionPopup being added with addWidget/addNew
Fixed some situations where WSuggestionPopup (a global widget) was added
as a widget. Because WSuggestionPopup's parent widget is always the
domRoot, only ownership should be transferred, so addChild should be
used.

Issue and patch provided by Bruce Toll.

We will revisit this for the next release, and consider adding
static_asserts and maybe a function like addNewChild.