Issue #7586: configurable trusted-proxies and original-ip-header
Instead of disabling Client-IP headers by default, let's just make
the header to use configurable, and have it default to X-Forwarded-For.

Also, instead of filtering out local IP addresses from X-Forwarded-For
and trying to get the client IP address that way, we'll just go back
to front until we occur an IP address that is not a trusted proxy.

Added trusted-proxy-config section to wt_config.xml, with:

 - original-ip-header
 - trusted-proxies

We will now trust headers coming from a reverse proxy
(e.g. X-Forwarded-* headers, SSL info headers) only if they are coming
from a trusted reverse proxy. This removes the caveat that Wt applications
should not be directly accessed if configured to be behind a reverse proxy.

Co-authored-by: Bruce Toll <btoll@dhsus.com>