Make addNew use addChild instead of addWidget for global widgets
WPopupWidget and such should not be added with addWidget to a
container widget (except for the domroot), since they are
global widgets. Instead of warning that this is something you're
not allowed to do, let's just make it use addChild instead,
and only transfer the ownership. That's likely what the user
intends to do here. This would prevent issues like #7669.