Added Jenkinsfiles and Dockerfiles
These run on anything with the 'docker' label on Jenkins.
This will make it use the new build machine. This machine
has 16 GiB RAM, 4 cores + HT, so we'll use 5 threads.

On FreeBSD there is currently only one thread, but we can scale
that up later.

Connection settings in DboFixture.h have been updated to match
the ones used in db.Jenkinsfile, instead of using sed to replace
them, which is what we used to do before.