Changed WFormModel to use string comparison instead of pointer comparison
It was a bad idea (probably an attempt at optimization) to use
string literals and pointer comparison for fields. Ideally,
the field would be identified by a std::string, but to avoid
having to change the API while at the same time partially
eliminating the issues caused by the pointer comparison,
we'll just use strcmp.