Issue #7749: ensure that request parameters are parsed only once
In WebController, CgiParser::parse is used once in handleRequest, and
once in requestDataReceived. requestDataReceived is called repeatedly
as uploads come in. This would cause a constant accumulation of duplicate
values for parameters, and cause a race condition if the value is being
updated while updateResourceProgress is being handled. This is most
noticeable when the upload is coming in very fast (e.g. on localhost).