WMediaPlayer: fixes for reload-is-new-session = false
When refreshing the page the JavaScript for jplayer was
loaded before jquery, causing JS errors. The jplayer JS
is loaded first because any JS file that is require()d
is inserted as a <script> tag when progressive bootstrap
is used, or when reload-is-new-session = false, and
the page is refreshed.

This change makes WMediaPlayer use requireJQuery, which
makes it load jquery just like any other JS file. This
is however not applicable if the user has already set
another version of jquery.

Also, on refresh the media source has to be set again,
hence the flags.test(RenderFlag::Full) && !media_.empty()