Updated date library to latest master (was version 2.4.1)
This is currently commit 97246a638a6d8f0269f4555c5e31106a86e3fd94

Howard Hinnant himself considers the API stable since 3.0.0, and
any future commits on master are merely fixes. So let's stick to
tip-of-master.

WDate.C: operator unsigned doesn't exist anymore for weekday, and
iso_encoding() was added, so we'll use that instead. iso_incoding()
already represents sunday as a 7 so we don't have to convert it.