WT-7965: Allow style requests after ajax mode is started
This commit permits style requests to be processed after ajax mode has
started, in order to avoid logging messages that are unnecessarily severe.

There are at least two valid cases where a style request with the same
wtd as an existing session can arrive for a bootstrap mode session that
has already transitioned to ajax mode:

1. Chromium-based browsers may re-send the "request=style" GET request
   the first time that the Developer Tools window is opened for a page.
   This request will currently receive a 403 response and log:
   "Unexpected GET request with wtd of existing Ajax session".

2. When Javascript is available, there are two client requests that
   arrive in an unspecified order: the script and the style requests.

   The style request typically arrives first -- and this works fine.

   If the script request arrives first, the session will transition
   to Ajax mode and the style request will not be needed -- the
   stylesheets will be added to the DOM via Javascript instead. However,
   when the style request subsequently arrives, it will receive a 403
   response and log: "Unexpected GET request with wtd of existing Ajax
   session".

The impact of returning a 403 seems to be harmless in both of the above
cases, but the log message makes it sound like a serious event.

This commit permits style requests to be processed after ajax mode has
started. This avoids logging the "Unexpected GET" message. Moreover,
the reply to the style request should match pre-4.4.0 behavior.

Co-authored-by: Korneel Dumon <korneel@emweb.be>