WT-7968: Revert delay of WWidget::load()
Reverts commit b67d3ccca94fa29e274870362de485d8a60204dc

The load() was delayed to after rendering to prevent issues when the
parent container of a widget is passed was passed into the constructor
of a widget due to load() (a virtual function) being called before
"this" was fully initialized.

Wt 4 makes it impossible to pass the parent container of a widget in the
constructor, so that delay is no longer necessary.

The delay could cause issues, like when widgets use require() in their
constructor. This could lead to JavaScript being executed out of order,
so there were JavaScript errors (e.g. when using WTextEdit).

Moving load() back to when the widget is added as a child would lead to
one other issue: WMenu would not correctly load the contents of the
first item when that item was added after WMenu was already added to a
container (the menu was already loaded). So when adding an item to an
empty menu, we need to not only set that to the currently selected item,
but also load its contents.