WT-7985: expose setHeaderData(int, std::any) in QueryModel
WAbstractItemModel has two setHeaderData functions:

One that is virtual and takes 4 arguments (section,
orientation, value, and role), and another that serves
as a shorthand and takes 2 arguments (section and value),
that calls the 4 argument function with orientation
set to Orientation::Horizontal and role set to ItemDataRole::Edit.

QueryModel overrides the 4-argument version, which causes the
2-argument function to be hidden.

Adding "using WAbstractTableModel::setHeaderData" to QueryModel makes it
so the 2-argument version is accessible.

The fourth argument is defaulted in WAbstractTableModel, so we don't
have to repeat it. The "using" should make it so that it is accessible
as well.