Resolved an issue with DBO_EXTERN_TEMPLATES and DBO_INSTANTIATE_TEMPLATES
The Wt::Dbo::query_result_traits<> template was declared and defined as 'struct'
in SqlTraits.h, but is declared as 'class' in the DBO_EXTERN_TEMPLATES
and DBO_INSTANTIATE_TEMPLATES macros. This inconsistency causes multiple
compiler warnings, such as C4099 on Visual Studio 2017/2019., whenever the
macros are used. Declaring as struct solves the issue.