Fixed '[c|tr].size is not a function' error when using jQuery 3.x.
While using jQuery 3.x with WApplication::requireJQuery, the following
error occurs if the server is restarted while there are still active
sessions, thus forcing the sessions to reload.

Wt internal error; code: undefined, description: tr.size is not a
function

This is caused by the use of the jQuery size() function, which was
deprecated in jQuery 1.8 and removed in jQuery 3.0. Replacing the size()
function with the length property as suggested in
https://api.jquery.com/size/ fixes the issue, while remaining full
compatibility with jQuery 1.0+.