WT-8022: fix ambiguous local_time and sys_time
Instead of "using namespace" I opted to make the namespaces explicit.
This eliminates any conflict between things that are defined in the
date library, and things defined in std::chrono since C++20 (since
C++20 std::chrono contains most of the date library)

We'll have to add the option to allow C++20's std::chrono support
instead of the date library at some point in the future.