Fix regression when appending to WString with args
In order to fix the case of WString("{1}").arg("").empty() the definition of
literal() was changed to consider a string not literal if it had any arguments
bound to it. This was a mistake since literal() is widely regarded in other
code to mean that it is not localized, regardless of whether there are any arguments.

Whenever a WString is appended to, it is first made literal by resolving the
key. If the string was not localized but had arguments it would however try to
resolve an empty key, fail and return "????".

The change to literal() was reverted, and now empty() simply checks if impl_ is
null or not. If impl_ is not null that means the string either is localized or has
arguments.

Added a test case to test the regression.