Custom model to represent database objects
This class inherits from FormModelBase and will be automatically
initialized with the values in the database table. The users needs
to start with defining which columns they want to add to the model.
This can be done in three ways:
- addDboField: to add them separately
- addDboFields: to add a list of Dbo columns
- addAllDboColumnsAsFields: to add all of the columns to the model
Once all of the Dbo fields have been specified, they need to call
initDboValues to pull data from the database into the model.