Custom view to represent database objects
This class inherits from WTemplateFormView and uses the custom Dbo
actions and model to initialize the view. When the custom model is
set, we will use the custom Dbo ViewAction to automatically generate
the default form delegates for the database columns added in the model.
The user is able to override the default form delegates by using
setFormDelegate.
Note: they will only be able to call this method before setting the model.

The user is also able to customize the widgets and validators of the
default form delegates, without having to specify an entire form
delegate. This can for example be useful for an integer column. The
default validator is a WIntValidator. Using the customizeValidator
function will allow them to set a range.