Custom model to keep track of database columns
In future commits I will be adding custom Dbo actions to be used
with Dbo's persist method to create a link between the custom
WFormModel and the database object. This base class will be used
by that custom WFormModel and those actions to keep track of
which database columns have been added to the model.