Use ccache on Jenkins
ccache is now installed in all of the Docker images, and used
through -DCMAKE_C(XX)_COMPILER_LAUNCHER=ccache, except for minver,
which uses an older version of CMake, where
-DCMAKE_C(XX)_COMPILER=/usr/lib/ccache/... is used.

Set the size of the cache to 20G, because of the many different builds
overwriting eachother when it was set to the default of 5G, causing many
cache misses.

The cache uses a bind mount instead of a volume, so it won't
accidentally be purged. The :z option is only relevant on systems
with SELinux but I included it for completeness. It sets the SELinux
label of that directory. :z indicates that the bind mount content is
shared among multiple containers.