Changes to HttpClientServerTest.C to reduce crashes
I don't think I have fully figured out how to make it not
crash sometimes, there may still be issues with
Wt::Http::Client itself but every time I try to understand
it I get a headache.

This change mainly makes sure that the destructor of Client
waits until the done() signal is emitted. This makes sure that
the client always outlives the signal.

Instead of inheriting from Http::Client we're making sure that
done_ is in the correct state by having Http::Client as a member.

Also, isDone() can be const so let's make it const.