WT-8838: unblock thread-directed signals
We block all signals, presumably because we want to be sure they're
handled by the main thread, and not some random worker thread. There are
however some special signals that are always directed at a specific
thread:

 - SIGBUS (Bus error (bad memory access))
 - SIGFPE (Floating point exception)
 - SIGILL (Illegal instruction)
 - SIGSEGV (Segmentation fault)

Those signals should not be blocked if we want signal handlers from
libraries like breakpad or crashpad to do their job.