WT-8610: segfault in WebController::generateNewSessionId
Since the UpdateLock was only taken after the session was removed from
the sessions_ map, there was a chance for a race condition with an incoming
request. When the request concerns a user login, the application wants
to change the session-id, but it can't since the session was already removed
from the map.
As a fix, removing the session is now inside the UpdateLock. I checked that
the UpdateLock is never taken after the WebController::mutex_ as this might
lead to deadlock.