WT-8622: Complete Dbo Forms documentation
Documentation for the Dbo Forms was already written when it was
first implemented, but it seems I missed a few things. This commit
focuses on completing the existing documentation and adding some
for parts that were missed.

Note that I used `\internal` in some places. When the INTERNAL_DOCS
configuration option is set to "NO" these sections will be omitted
from the generated documentation.