WT-7418: Split WBootstrapTheme into WBootstrap2Theme and WBootstrap3Theme
WBootstrapTheme remains, but is deprecated. It acts as a bridge to the
concrete implementation: WBootstrap2Theme or WBootstrap3Theme.

This allows the interface of different Bootstrap theme implementations
to be more different, and makes way for the introduction of
WBootstrap4Theme and WBootstrap5Theme.

The files specific to a Bootstrap theme have also been moved into their
own directory.