WT-9230: make 3-argument version of QueryModel::setHeaderData accessible again
When adding using WAbstractTableModel::setHeaderData, I mistakenly also
removed the default value for the role argument. This made it so that
QueryModel::setHeaderData couldn't be used with 3 arguments anymore.

Default arguments on virtual functions are a bad idea in general, and
mixing overloading and inheritance can also be a world of pain, but
we're stuck with it now.