WT-7418: changed the nameless elementRole "1" into ToggleButtonRole
We were using this role, but didn't name it anything. It appears that
we're only using this for WAbstractToggleButton, so let's call it
ToggleButtonRole.

We can't use ToggleButton here, since that'll conflict with the
ToggleButton used in WTreeView.C.

I'm not sure why we don't just use MainElement here, though?