WT-9204: move WApplication-changing code from Bootstrap theme ctor to init function
Themes are assigned using a shared pointer, with the idea being that
the theme is stateless once configured. Themes could theoretically be
shared between instances of WApplication. However, in practice, the
Bootstrap themes had to always be instantiated in the constructor of
WApplication, since the themes affect the WApplication.

To make the themes properly stateless the WApplication-changing code was
moved to an init function that is called when WApplication::setTheme()
is called.

Co-authored-by: Korneel Dumon <korneel@emweb.be>