WT-9363: workaround for Boost Spirit ≥ 1.76.0
Since Boost 1.76.0 there is a regression in the outputting of numbers:
https://github.com/boostorg/spirit/issues/688

Wt depends on this to serialize doubles to JavaScript.

Until Boost incorporates a fix for this issue we will disable the use of
Spirit to serialize doubles and use sprintf instead for versions of
Boost starting from 1.76.0.

Hopefully it will be fixed in 1.78.0, at which point we can update the
check.