WT-9391: Add WServer::docRoot()
Sometimes we may want to use docRoot() outside of the context of a
WApplication. It would be nice to have a WServer::docRoot() that
provides the value that was given to the --docroot command line argument
in wthttp. A caveat is that this would only be usable with wthttp, but
that's hardly an issue, since that is the recommended way of deploying
Wt anyway (I would actually actively discourage the use of wtfcgi or
wtisapi).