WT-9353: auth1 example hangs on oauth login
Can't run onOAuthDone() directly in the OAuthRedirectEndpoint, since
it might cause the resource to be deleted. The application then deadlocks
in WResource::beingDeleted which will wait for the resource handling to
complete.
This happens in AuthWidget which deletes the OAuthProcess when the login
changes.

As a solution, the 'load' signal from the application is used (this will
trigger unsuspended_). This is more in line with the workflow using a
popup, which will trigger the 'redirected_' signal.