WT-9328: improve tooltip positioning
Made a separate method to position tooltips, this way I don't have to worry
about breaking the positioning of other elements and it allows me to use
fixed position (relative to viewport). This makes the calculation easier.

For vertical positioning the preferred position is now: below, above, shift
up from below, fill page (will overflow).
For horizontal position: right, left, side with the most space (will overflow).

Fixed position will require the tooltip itself to scroll if it overflows.
This is done by setting max-width/height. The mousleave handler is moved to
outerDiv because otherwise the tooltip disappears if the mouse is on the
scrollbar.