WT-9571: fix false positive -Wfree-nonheap-object warning
Doing it this way ensures that GCC understands that the delete this in
decref() is never actually called for end_link.

Instead of incrementing the reference count of end_link so that it
doesn't get deleted in unlink(), we instead add a bool parameter to
unlink that determines whether the unlink should cause a decref(). By
default it always does so, but end_link, being stack-allocated, is a
special case.

unlink is probably inlined by the compiler, so this extra if may
actually marginally improve performance?