WT-8760: always show invalid login
The check for email verification was only done when login was attempted
(in FormBaseModel::loginUser) rather than when the form was validated.
This meant that if the user did not remember their password it was not
possible to resend the email verification.
I also keep the check in FormBaseModel since it is also called from the
RegistrationModel.