WT-9035: fix self-closing tags in config <head-matter>
When a script tag was in the <head-matter> config, the xml parser would
automatically make it self-closing if it had no content (which happens
for example when the 'src' attribute is set). This fix will add a closing
tag.

A helper function from xml message resource bundles is reused for this.