WT-9562: fix for JWt
std::ostringstream gets turned into java.io.StringWriter, and
Message::write gets turned into javax.mail.Message.writeTo. That
function however takes a java.io.OutputStream.

Using WT_BAOSTREAM we can use a std::ostringstream for Wt and the
non-existing std::baostream for JWt (which gets turned into
java.io.ByteArrayOutputStream).

We'll not do the imbue bit for JWt. We only do it to be safe in C++,
because of the global locale possibly breaking things.

The Doxyfile change is for whenever we'd put WT_BAOSTREAM in a header,
than it is still displayed as std::ostringstream in the generated
documentation.