WT-10287: wthttp: acknowledge WebSockets close opcode
According to RFC 6455 5.5.1: If an endpoint receives a Close frame and
did not previously send a Close frame, the endpoint MUST send a Close
frame in response.

Previously, Wt would not send the Close frame in response, causing
Firefox to view the connection as faulty and trigger its exponential
backoff.