WT-10589: Add public getters to model and view
The user may want to do something with the `item_` after they've saved
it in the view. To allow this, we need to make the `item_` as well as
the `model_` publicly accessible.

Now the user could react to the `saved()` signal from a specialized
`FormView` class by for example saying "Item #ID has been successfully
saved".