WT-10488: the onload handler should also check http status code
The onload handler would not check the status code to see if it should
pass true or false to handleResponse. Since onload is also called when
there is an HTTP error, it should also check this.

However, in a quick test, the onreadystatechange handler was always run
before the onload handler. This is probably the reason this was never a
problem.