WT-10488: keepAlive was no longer queued in case of request error
When a website is idle and has server-push enabled, a `commError` caused
by 504 (gateway timeout) will stop keepAlive signals from being sent. The
commError is never reset because the long-polling events are never
"successfull" since they are aborted after a timeout. For an idle page,
this can cause the session to be destroyed after a while.

Fixed by not considering 504 error on polling requests as `commError` (so
keepAlive is still sent).