WT-8801: Adopt a policy for the minimum version of main dependencies
While solid backwards compatibility is one of Wt's goals, it's not
realistic to have to continue to support very old versions of our main
dependencies.

By adopting this policy we can make it predictable for our users, and
will avoid any discussion about whether or not we will drop support for
certain versions.