WT-10600: added minimal.Jenkinsfile for macOS
We're doing shared and static builds using a matrix so we get a nice
overview in Blue Ocean, and see all of the test results nicely
separated.

We're using agent none outside of the matrix and agent { label 'mac' }
inside of the matrix, so that the branches don't run concurrently
on the same executor.