WT-10834: fix dbo_test46 for MariaDB and SQL Server
The test was failing with the following assertion failure:
ASSERTION FAILURE:
- file   : DboTest.C
- line   : 3482
- message: critical check std::get<0>(*it2) has failed

This is due to the fact that the result of the query was being
ordered by a.id, and when PostgreSQL orders a column it puts null
values last, whereas MariaDB and SQL Server put null values first.

A simple fix is to order by b.id, because we know that that column
will not have null values, so it will be consistent across backends.