WT-9076: date edit crashes on first render when using widgetset
Wt seems to render in two phases, first only the visible stuff, then the
rest.
For a non-widgetset application, a Script type response will trigger
rendering the visible stuff (in serveMainAjax) and a call to 'load' which
will trigger the second phase.
For widgetset applications however, the Script type response will itself
generate a call to 'load' which triggers rendering the visible stuff
(again by serveMainAjax). In this case, the 'load' signal is not triggered
again, so no second rendering phase takes place.

The crash happens because the second rendering phase did not take place.
I fixed it by triggering another 'none' signal for widgetset apps.

In addition, the twoPhaseThreshold was not taken into account in
serveMainAjax(), so I added it there (and refactored a bit for reuse).