WT-10941: change filetostring.cmake to split into string literals
Our new minifier (terser) would create minified JavaScript files that
had fewer line breaks than our old minifier
(an old version of Google closure compiler). MSVC would return a C2026
error when it encountered these very long string literals.

Older versions of MSVC would limit string constants to 64K, but this is
not an issue anymore, so we can change filetostring.cmake to instead
split the file up into multiple string literals, and just put all of it
into a single variable.

All files that used these string literals were changed accordingly,
removing the bits where we're concatenating the vector of strings, and
changing the variable to no longer include an index suffix.

I also changed the formatting on file_to_string in src/CMakeLists.txt,
changing it to a function, and added the filetostring.cmake script as
a dependency so that updates to that file will be picked up when
rebuilding.