Use WT_SOURCE_DIR to make sure add_subdirectory(wt) works properly
In order to be able to properly build and install Wt when it is used as
a subdirectory of another project we'll use WT_SOURCE_DIR consistently.

Notably, when add_subdirectory(wt) was used and the wt/doc target was
built, asciidoc would fail because the path to the .doc file would be
wrong.