WT-10966: Switch to PNPM
We need a more recent version of Node.js to install dprint
than the one that comes with Ubuntu 22.04. PNPM has the
pnpm env feature that conveniently allows us to install
a different version of Node.js.

As an added benefit, pnpm is supposed to be faster and more
space efficient than NPM.