Add WT_DEPRECATED macro
This macro translates to [[deprecated(details)]] when Wt is being used
as a library, but disables the warning for any internal uses. This is so
we can mark parts of Wt as deprecated without getting a bunch of
warnings because of internal uses that only exist in order to maintain
backwards compatibility.