WT-9350: update Bootstrap 5 theme and make it customizable
Instead of including a copy of the Bootstrap 5 theme sources, we'll use
PNPM to build it instead.

All of the Wt-specific additions were translated to SCSS and put in
their own folder. To organize it a bit, the SCSS files are now split up
according to which widget they are related to.

The WDialog template needed some fixing for the update to Bootstrap
5.2.3 (from 5.1.3)