WT-10364: BS5 theme: made WPanel collapsible again
Instead of using an accordion for collapsible WPanel, and
a card for a non-collapsible WPanel, we now use a card for both again.

This makes the look of collapsible and non-collapsible panels more
consistent.

To make this work, we added some extra CSS specifically for collapsible
cards in Wt (.card.Wt-panel.Wt-collapsible). The collapse button's image is
derived from the accordion button.

We're now using a WIconPair again for a collapsible WPanel, reverting a
lot of the code in WPanel to how it was before. Instead of using the
ugly icons, the theme overrides these with a one pixel transparent gif
and uses CSS to show a nicer icon, which is also animated.

WTheme::setDataTarget is no longer necessary, and has been removed.

WTheme::panelCollapseIconSide has been added to determine whether the
collapse icon should be added at the beginning of the titleBarWidget()
or to the end.

WT-10395 was also addressed: panels are automatically made
non-collapsible if the title bar is disabled.